ISSUE #2160: Improve import flow and get rid of unnecessary cron complexity
