--- src/Application/Import/RunStravaImportAndBuildAppCronAction.php
@@ -1,142 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Application\Import;
-
-use App\Application\AppUrl;
-use App\Application\Build\RunBuild\RunBuild;
-use App\Application\Import\RunStravaImport\RunStravaImport;
-use App\Domain\Activity\ActivityId;
-use App\Domain\Activity\ActivityIds;
-use App\Domain\Activity\ActivityRepository;
-use App\Domain\Import\ImportMode;
-use App\Domain\Integration\Notification\SendNotification\SendNotification;
-use App\Domain\Strava\Webhook\WebhookAspectType;
-use App\Domain\Strava\Webhook\WebhookEventRepository;
-use App\Infrastructure\CQRS\Command\Bus\CommandBus;
-use App\Infrastructure\Daemon\Cron\RunnableCronAction;
-use App\Infrastructure\DependencyInjection\Mutex\WithMutex;
-use App\Infrastructure\Doctrine\Migrations\MigrationRunner;
-use App\Infrastructure\Mutex\LockIsAlreadyAcquired;
-use App\Infrastructure\Mutex\LockName;
-use App\Infrastructure\Mutex\Mutex;
-use App\Infrastructure\Time\ResourceUsage\ResourceUsage;
-use Symfony\Component\Console\Style\SymfonyStyle;
-
-#[WithMutex(lockName: LockName::IMPORT_DATA_OR_BUILD_APP)]
-final readonly class RunStravaImportAndBuildAppCronAction implements RunnableCronAction
-{
-    public function __construct(
-        private CommandBus $commandBus,
-        private WebhookEventRepository $webhookEventRepository,
-        private ActivityRepository $activityRepository,
-        private ResourceUsage $resourceUsage,
-        private AppUrl $appUrl,
-        private Mutex $mutex,
-        private MigrationRunner $migrationRunner,
-        private ImportMode $importMode,
-    ) {
-    }
-
-    public function getId(): string
-    {
-        return 'importDataAndBuildApp';
-    }
-
-    public function supportsConfiguredImportMode(): bool
-    {
-        return $this->importMode->isStravaApi();
-    }
-
-    public function requiresDatabaseSchemaToBeUpdated(): bool
-    {
-        return false;
-    }
-
-    public function getMutexTtl(): int
-    {
-        return 1800;
-    }
-
-    public function run(SymfonyStyle $output): void
-    {
-        $this->migrationRunner->run($output);
-        $this->mutex->acquireLock('importDataAndBuildAppCronAction');
-
-        $this->doRun(
-            output: $output,
-            restrictToActivityIds: null
-        );
-
-        $this->mutex->releaseLock();
-    }
-
-    public function runForWebhooks(SymfonyStyle $output): void
-    {
-        $this->migrationRunner->run($output);
-        try {
-            $this->mutex->acquireLock('processWebhooks');
-        } catch (LockIsAlreadyAcquired) {
-            // Another process is importing data, postpone webhook processing.
-            return;
-        }
-
-        if (!$webhookEvents = $this->webhookEventRepository->grab()) {
-            // No webhooks to process.
-            $output->writeln('No webhook events left to process...');
-
-            return;
-        }
-
-        $activityIdsToDelete = ActivityIds::empty();
-        $createOrUpdateActivityIds = ActivityIds::empty();
-        foreach ($webhookEvents as $webhookEvent) {
-            $activityId = ActivityId::fromUnprefixed($webhookEvent->getObjectId());
-            if (WebhookAspectType::DELETE === $webhookEvent->getAspectType()) {
-                $activityIdsToDelete->add($activityId);
-            } else {
-                $createOrUpdateActivityIds->add($activityId);
-            }
-        }
-
-        if (!$activityIdsToDelete->isEmpty()) {
-            $this->activityRepository->markActivitiesForDeletion($activityIdsToDelete);
-        }
-
-        $this->doRun(
-            output: $output,
-            restrictToActivityIds: $createOrUpdateActivityIds
-        );
-
-        $this->mutex->releaseLock();
-    }
-
-    private function doRun(
-        SymfonyStyle $output,
-        ?ActivityIds $restrictToActivityIds,
-    ): void {
-        $this->resourceUsage->startTimer();
-
-        $this->commandBus->dispatch(new RunStravaImport(
-            output: $output,
-            restrictToActivityIds: $restrictToActivityIds,
-        ));
-        $this->commandBus->dispatch(new RunBuild(
-            output: $output,
-        ));
-
-        $this->resourceUsage->stopTimer();
-        $this->commandBus->dispatch(new SendNotification(
-            title: 'Build successful',
-            message: sprintf('New import and build of your stats was successful in %ss', $this->resourceUsage->getRunTimeInSeconds()),
-            tags: ['+1'],
-            actionUrl: $this->appUrl
-        ));
-
-        $output->writeln(sprintf(
-            '<info>%s</info>',
-            $this->resourceUsage->format(),
-        ));
-    }
-}
