+++ src/Console/Daemon/RunFileImportAndBuildAppConsoleCommand.php
@@ -6,7 +6,8 @@ namespace App\Console\Daemon;
 
 use App\Application\AppUrl;
 use App\Application\Build\RunBuild\RunBuild;
-use App\Application\Import\RunFileImport\RunFileImport;
+use App\Application\Import\CalculateActivityMetrics\CalculateActivityMetrics;
+use App\Application\Import\FileImport\ImportActivityFiles;
 use App\Domain\Activity\ActivityIdRepository;
 use App\Domain\Import\WatchDirectory;
 use App\Domain\Integration\Notification\SendNotification\SendNotification;
@@ -14,25 +15,37 @@ use App\Infrastructure\CQRS\Command\Bus\CommandBus;
 use App\Infrastructure\DependencyInjection\Mutex\WithMutex;
 use App\Infrastructure\Doctrine\Migrations\MigrationRunner;
 use App\Infrastructure\Exception\EntityNotFound;
+use App\Infrastructure\FileSystem\PermissionChecker;
 use App\Infrastructure\KeyValue\Key;
 use App\Infrastructure\KeyValue\KeyValue;
 use App\Infrastructure\KeyValue\KeyValueStore;
 use App\Infrastructure\KeyValue\Value;
+use App\Infrastructure\Logging\LoggableConsoleOutput;
 use App\Infrastructure\Mutex\LockIsAlreadyAcquired;
 use App\Infrastructure\Mutex\LockName;
 use App\Infrastructure\Mutex\Mutex;
 use App\Infrastructure\Time\Clock\Clock;
 use App\Infrastructure\Time\ResourceUsage\ResourceUsage;
+use Doctrine\DBAL\Connection;
+use League\Flysystem\UnableToCreateDirectory;
+use League\Flysystem\UnableToWriteFile;
+use Monolog\Attribute\WithMonologChannel;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
+#[WithMonologChannel('console-output')]
 #[WithMutex(lockName: LockName::IMPORT_DATA_OR_BUILD_APP)]
-#[AsCommand(name: 'app:cron:run-file-import', description: 'Run file import')]
+#[AsCommand(name: RunFileImportAndBuildAppConsoleCommand::NAME, description: 'Run file import')]
 final class RunFileImportAndBuildAppConsoleCommand extends Command
 {
+    public const string NAME = 'app:cron:run-file-import';
+    public const string FORCE_OPTION = 'force';
+
     public function __construct(
         private readonly CommandBus $commandBus,
         private readonly ActivityIdRepository $activityIdRepository,
@@ -43,14 +56,27 @@ final class RunFileImportAndBuildAppConsoleCommand extends Command
         private readonly AppUrl $appUrl,
         private readonly Clock $clock,
         private readonly KeyValueStore $keyValueStore,
+        private readonly PermissionChecker $fileSystemPermissionChecker,
+        private readonly Connection $connection,
+        private readonly LoggerInterface $logger,
     ) {
         parent::__construct();
     }
 
+    protected function configure(): void
+    {
+        $this->addOption(RunStravaImportAndBuildAppConsoleCommand::SKIP_IMPORT_OPTION, null, InputOption::VALUE_NONE);
+        $this->addOption(RunStravaImportAndBuildAppConsoleCommand::SKIP_BUILD_OPTION, null, InputOption::VALUE_NONE);
+        $this->addOption(self::FORCE_OPTION, null, InputOption::VALUE_NONE);
+    }
+
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $output = new SymfonyStyle($input, $output);
-        $this->migrationRunner->run($output);
+        $output = new SymfonyStyle($input, new LoggableConsoleOutput($output, $this->logger));
+        $skipImport = $input->getOption(RunStravaImportAndBuildAppConsoleCommand::SKIP_IMPORT_OPTION);
+        if (!$skipImport) {
+            $this->migrationRunner->run($output);
+        }
 
         $today = $this->clock->getCurrentDateTimeImmutable()->format('Y-m-d');
         $watchDirectoryHasFiles = $this->watchDirectory->hasFilesThatCanBeProcessed();
@@ -61,7 +87,7 @@ final class RunFileImportAndBuildAppConsoleCommand extends Command
             $alreadyBuiltToday = false;
         }
 
-        if (!$watchDirectoryHasFiles && $alreadyBuiltToday) {
+        if (!$skipImport && !$watchDirectoryHasFiles && $alreadyBuiltToday) {
             $output->writeln('No files left to process...');
 
             return Command::SUCCESS;
@@ -78,29 +104,46 @@ final class RunFileImportAndBuildAppConsoleCommand extends Command
             return Command::SUCCESS;
         }
 
-        if ($watchDirectoryHasFiles) {
-            $this->commandBus->dispatch(new RunFileImport(
-                output: $output,
+        if (!$skipImport && $watchDirectoryHasFiles) {
+            try {
+                $this->fileSystemPermissionChecker->ensureWriteAccess();
+            } catch (UnableToWriteFile|UnableToCreateDirectory) {
+                $output->writeln('<error>Make sure the container has write permissions to "storage/database" and "storage/files" on the host system</error>');
+
+                return Command::SUCCESS;
+            }
+
+            $this->commandBus->dispatch(new ImportActivityFiles($output));
+            $this->commandBus->dispatch(new CalculateActivityMetrics($output));
+
+            $this->connection->executeStatement('VACUUM');
+            $output->writeln('Database got vacuumed 🧹');
+        }
+
+        if (!$input->getOption(RunStravaImportAndBuildAppConsoleCommand::SKIP_BUILD_OPTION)) {
+            if ($this->activityIdRepository->count() <= 0) {
+                $this->mutex->releaseLock();
+                $output->writeln('<error>Wait until at least one activity has been imported before building the app</error>');
+
+                return Command::SUCCESS;
+            }
+
+            try {
+                $this->commandBus->dispatch(new RunBuild(
+                    output: $output,
+                ));
+            } catch (\Throwable $e) {
+                $this->logger->error($e->getMessage());
+                throw $e;
+            }
+
+            $this->keyValueStore->save(KeyValue::fromState(
+                key: Key::APP_LAST_BUILT_ON,
+                value: Value::fromString($today),
             ));
         }
 
-        if ($this->activityIdRepository->count() <= 0) {
-            $this->mutex->releaseLock();
-            $output->writeln('<error>Wait until at least one activity has been imported before building the app</error>');
-
-            return Command::SUCCESS;
-        }
-
-        $this->commandBus->dispatch(new RunBuild(
-            output: $output,
-        ));
-
         $this->mutex->releaseLock();
-
-        $this->keyValueStore->save(KeyValue::fromState(
-            key: Key::APP_LAST_BUILT_ON,
-            value: Value::fromString($today),
-        ));
 
         $this->resourceUsage->stopTimer();
         $this->commandBus->dispatch(new SendNotification(
