--- tests/Application/Import/RunStravaImport/RunStravaImportCommandHandlerTest.php
@@ -1,76 +1,0 @@
-<?php
-
-namespace App\Tests\Application\Import\RunStravaImport;
-
-use App\Application\Import\RunStravaImport\RunStravaImport;
-use App\Application\Import\RunStravaImport\RunStravaImportCommandHandler;
-use App\Domain\Strava\Strava;
-use App\Infrastructure\CQRS\Command\Bus\CommandBus;
-use App\Infrastructure\Serialization\Json;
-use App\Tests\ContainerTestCase;
-use App\Tests\Infrastructure\CQRS\Command\Bus\SpyCommandBus;
-use App\Tests\Infrastructure\FileSystem\SuccessfulPermissionChecker;
-use App\Tests\Infrastructure\FileSystem\UnwritablePermissionChecker;
-use App\Tests\SpyOutput;
-use Doctrine\DBAL\Connection;
-use PHPUnit\Framework\MockObject\MockObject;
-use Spatie\Snapshots\MatchesSnapshots;
-use Symfony\Component\Console\Input\StringInput;
-use Symfony\Component\Console\Style\SymfonyStyle;
-
-class RunStravaImportCommandHandlerTest extends ContainerTestCase
-{
-    use MatchesSnapshots;
-
-    private RunStravaImportCommandHandler $importStravaDataCommandHandler;
-    private CommandBus $commandBus;
-    private MockObject $connection;
-
-    public function testHandle(): void
-    {
-        $this->connection
-            ->expects($this->once())
-            ->method('executeStatement')
-            ->with('VACUUM');
-
-        $output = new SpyOutput();
-        $this->importStravaDataCommandHandler->handle(new RunStravaImport(
-            output: new SymfonyStyle(new StringInput('input'), $output),
-            restrictToActivityIds: null,
-        ));
-        $this->assertMatchesTextSnapshot(str_replace(' ', '', $output));
-        $this->assertMatchesJsonSnapshot(Json::encode($this->commandBus->getDispatchedCommands()));
-    }
-
-    public function testHandleWithInsufficientPermissions(): void
-    {
-        $this->connection
-            ->expects($this->never())
-            ->method('executeStatement');
-
-        $this->importStravaDataCommandHandler = new RunStravaImportCommandHandler(
-            $this->getContainer()->get(Strava::class),
-            $this->commandBus = new SpyCommandBus(),
-            new UnwritablePermissionChecker(),
-            $this->connection,
-        );
-
-        $output = new SpyOutput();
-        $this->importStravaDataCommandHandler->handle(new RunStravaImport(
-            output: new SymfonyStyle(new StringInput('input'), $output),
-            restrictToActivityIds: null,
-        ));
-        $this->assertMatchesTextSnapshot(str_replace(' ', '', $output));
-    }
-
-    #[\Override]
-    protected function setUp(): void
-    {
-        $this->importStravaDataCommandHandler = new RunStravaImportCommandHandler(
-            $this->getContainer()->get(Strava::class),
-            $this->commandBus = new SpyCommandBus(),
-            new SuccessfulPermissionChecker(),
-            $this->connection = $this->createMock(Connection::class),
-        );
-    }
-}
