--- tests/Infrastructure/Daemon/Cron/FakeRunnableCronAction.php
@@ -1,41 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Infrastructure\Daemon\Cron;
-
-use App\Infrastructure\Daemon\Cron\RunnableCronAction;
-use Symfony\Component\Console\Style\SymfonyStyle;
-
-final readonly class FakeRunnableCronAction implements RunnableCronAction
-{
-    public function __construct(
-        private bool $supportsConfiguredImportMode = true,
-    ) {
-    }
-
-    public function getId(): string
-    {
-        return 'fake';
-    }
-
-    public function supportsConfiguredImportMode(): bool
-    {
-        return $this->supportsConfiguredImportMode;
-    }
-
-    public function getMutexTtl(): int
-    {
-        return 60;
-    }
-
-    public function requiresDatabaseSchemaToBeUpdated(): bool
-    {
-        return true;
-    }
-
-    public function run(SymfonyStyle $output): void
-    {
-        $output->writeln('RunnableCronAction fake has been called');
-    }
-}
